ExpandableRecyclerView moved to external module
