+++ app/src/main/java/com/omega_r/omegarecyclerview/expandable_example/core/SimpleData.java
@@ -2,7 +2,8 @@ package com.omega_r.omegarecyclerview.expandable_example.core;
 
 import androidx.annotation.Nullable;
 
-import com.omega_r.libs.omegarecyclerview.expandable_recycler_view.data.GroupProvider;
+
+import com.omega_r.libs.omegarecyclerview_expandable.data.GroupProvider;
 
 import java.util.Arrays;
 import java.util.List;
