+++ app/src/main/java/com/omega_r/omegarecyclerview/expandable_example/support_sticky/DefaultStickyExpandableActivity.java
@@ -1,6 +1,6 @@
 package com.omega_r.omegarecyclerview.expandable_example.support_sticky;
 
-import com.omega_r.libs.omegarecyclerview.expandable_recycler_view.OmegaExpandableRecyclerView;
+import com.omega_r.libs.omegarecyclerview_expandable.OmegaExpandableRecyclerView;
 import com.omega_r.omegarecyclerview.expandable_example.core.ExpandableActivity;
 
 import omega.com.annotations.OmegaActivity;
