--- omegarecyclerview/src/main/java/com/omega_r/libs/omegarecyclerview/expandable_recycler_view/animation/OnAnimationEndListener.java
@@ -1,6 +1,0 @@
-package com.omega_r.libs.omegarecyclerview.expandable_recycler_view.animation;
-
-public interface OnAnimationEndListener {
-    void onAnimationEnd();
-}
-
