--- omegarecyclerview/src/main/java/com/omega_r/libs/omegarecyclerview/expandable_recycler_view/data/ExpandableType.java
@@ -1,5 +1,0 @@
-package com.omega_r.libs.omegarecyclerview.expandable_recycler_view.data;
-
-public enum ExpandableType {
-    GROUP, CHILD
-}
