+++ omegarecyclerview_expandable/src/main/java/com/omega_r/libs/omegarecyclerview_expandable/animation/OnAnimationEndListener.java
@@ -1,0 +1,6 @@
+package com.omega_r.libs.omegarecyclerview_expandable.animation;
+
+public interface OnAnimationEndListener {
+    void onAnimationEnd();
+}
+
