+++ omegarecyclerview_expandable/src/main/java/com/omega_r/libs/omegarecyclerview_expandable/data/ExpandableType.java
@@ -1,0 +1,5 @@
+package com.omega_r.libs.omegarecyclerview_expandable.data;
+
+public enum ExpandableType {
+    GROUP, CHILD
+}
