+++ omegarecyclerview_expandable/src/main/java/com/omega_r/libs/omegarecyclerview_expandable/data/UniqueIdProvider.java
@@ -1,0 +1,5 @@
+package com.omega_r.libs.omegarecyclerview_expandable.data;
+
+public interface UniqueIdProvider {
+    long provideUniqueId();
+}
