+++ omegarecyclerview_expandable/src/test/java/com/omega_r/libs/omegarecyclerview_expandable/ExampleUnitTest.java
@@ -1,0 +1,17 @@
+package com.omega_r.libs.omegarecyclerview_expandable;
+
+import org.junit.Test;
+
+import static org.junit.Assert.*;
+
+/**
+ * Example local unit test, which will execute on the development machine (host).
+ *
+ * @see <a href="http://d.android.com/tools/testing">Testing documentation</a>
+ */
+public class ExampleUnitTest {
+    @Test
+    public void addition_isCorrect() {
+        assertEquals(4, 2 + 2);
+    }
+}
